Use mlflow.azureml.build_image()
