--- MLproject
@@ -1,12 +1,0 @@
-name: automl
-conda_env: conda.yml
-
-entry_points:
-    preprocess:
-        command: "python -c 'from automl.preprocess import preprocess; preprocess()'"
-    train_h2o:
-        command: "python -c 'from automl.h2o import train_h2o; train_h2o()'"
-    train_autogluon:
-        command: "python -c 'from automl.autogluon import train_autogluon; train_autogluon()'"
-    test:
-        command: "pytest automl/test.py"
