+++ Makefile
@@ -1,8 +1,11 @@
+include .env
+export
+
 # SHELL := /bin/bash
-SRC = automl/*.py
+SRC = */*.py
 DATA = data/processed/train.csv data/test/dftest.csv
 PREP_MODEL = models/prep.model
-MLFLOW_RUN = mlflow run .
+MLFLOW_RUN = mlflow run
 
 all: train serve
 train: train_autogluon
@@ -10,18 +13,18 @@ tmp/run_env.sh: $(SRC) $(DATA)
 	$(MAKE) train
 
 train_h2o: $(SRC) $(DATA) $(PREP_MODEL)
-	$(MLFLOW_RUN) -e train_h2o
+	$(MLFLOW_RUN) h2o_mlflow -e train_h2o
 
 train_autogluon: $(SRC) $(DATA) $(PREP_MODEL)
-	$(MLFLOW_RUN) -e train_autogluon
+	$(MLFLOW_RUN) autogluon_mlflow -e train_autogluon
 
-preproc $(DATA) $(PREP_MODEL): automl/preprocess.py
-	$(MLFLOW_RUN) -e preprocess
+preproc $(DATA) $(PREP_MODEL): common/preprocess.py
+	$(MLFLOW_RUN) common -e preprocess
 
 devenv: conda*.yml
 	conda env update -f conda.yml
 	conda env update -f conda-dev.yml
-	python --version 
+	python --version
 	gcc --version
 	java -version
