+++ autogluon_mlflow/conda.yml
@@ -7,7 +7,6 @@ dependencies:
   - bokeh
   - joblib
   - loguru
-  # - openjdk
   - gcc_linux-64
   - mlflow=1.11.0
   - pandas
@@ -15,7 +14,6 @@ dependencies:
   - python-dotenv
   - pip=20.1.1
   - pip:
-      # - h2o
       # - mxnet_cu101<2.0.0
       - mxnet<2.0.0
       - autogluon.tabular==0.0.15b20201105
