+++ common/conda.yml
@@ -1,0 +1,13 @@
+channels:
+  - defaults
+  - anaconda
+  - conda-forge
+dependencies:
+  - python=3.6.10 # match AzureML base image
+  - joblib
+  - loguru
+  - mlflow=1.11.0
+  - pandas
+  - scikit-learn
+  - python-dotenv
+  - pytest
