+++ common/test.py
@@ -10,7 +10,7 @@ import pandas as pd
 import requests
 from loguru import logger
 
-from automl.config import TEST_CSV
+from common.config import TEST_CSV
 
 
 def read_dftest():
@@ -28,13 +28,12 @@ def reload_mlflow_predict(dftest):
 
 
 def test_api():
+    scoring_uri = os.getenv("SCORING_URI", "http://127.0.0.1:5000/invocations")
     dftest = read_dftest()
 
     data = dftest.to_json(orient="split", index=False)
     res = requests.post(
-        "http://127.0.0.1:5000/invocations",
-        data=data,
-        headers={"Content-type": "application/json"},
+        scoring_uri, data=data, headers={"Content-type": "application/json"},
     )
 
     preds = pd.DataFrame(
