+++ h2o_mlflow/conda.yml
@@ -1,0 +1,17 @@
+channels:
+  - defaults
+  - anaconda
+  - conda-forge
+dependencies:
+  - python=3.6.10 # match AzureML base image
+  - joblib
+  - loguru
+  - openjdk
+  - mlflow=1.11.0
+  - pandas
+  - scikit-learn
+  - python-dotenv
+  - pytest
+  - pip=20.1.1
+  - pip:
+      - h2o
