+++ scorer/preproc_base.py
@@ -4,10 +4,7 @@ from typing import Union
 import joblib
 import pandas as pd
 from loguru import logger
-from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import MinMaxScaler
-
-from automl.config import DATA_URI, MODEL_DIR, PROCESSED_DATA_DIR, TEST_CSV
 
 
 class Preproc:
@@ -35,28 +32,3 @@ class Preproc:
     def save_model(self, model_path: Union[Path, str]) -> str:
         joblib.dump(self, model_path)
         return model_path
-
-
-def preprocess(csv=DATA_URI):
-    # %% 実験データのダウンロード
-    data = pd.read_csv(csv)
-    logger.debug("Downloaded data")
-
-    # %% 学習・テストデータの分割
-    dftrain: pd.DataFrame
-    dftest: pd.DataFrame
-    dftrain, dftest = train_test_split(data)
-    logger.info(f"train_test_split {data.shape} -> {dftrain.shape} + {dftest.shape}")
-
-    # Save test dataframe for later tests
-    Path("../data").mkdir(exist_ok=True)
-    dftest.to_csv(TEST_CSV, index=False)
-
-    # %% データ前処理
-    prep = Preproc()
-    train = prep.fit_transform(dftrain)
-    pre_model = prep.save_model(MODEL_DIR / "prep.model")
-    test = prep.transform(dftest)
-    train.to_csv(PROCESSED_DATA_DIR / "train.csv", index=False)
-    test.to_csv(PROCESSED_DATA_DIR / "test.csv", index=False)
-    return train, test, pre_model
