+++ scripts/clean_code.sh
@@ -1,8 +1,6 @@
 set -ex
 shopt -s globstar
 
-cd automl
-
 autoflake --in-place --remove-all-unused-imports --remove-unused-variables \
     **/*.py
